whitespace fixes
A day may come when new CSS is needed, and I submit a pull request with no whitepsace issues, but it is not this day. An hour of joy and celebration, when I successfully catch my mistakes and the age of whitespace comments comes crashing down! But it is not this day! This day I submit more fix commits, and hang my head in shame.