updates logic for showing the cart markup
this change is required because when the cart is empty the surrounding
span is not generated but then this leaves missing markup. In the event
that we want to increment the cart count using js and the current cart
has no items in it then the span is not rendered and then therefore not
incrementable.